refactor($default-theme): simplify & reorganize
