+++ packages/@vuepress/theme-default/components/Navbar.vue
@@ -92,7 +92,6 @@ $navbar-horizontal-padding = 1.5rem
 .navbar
   padding $navbar-vertical-padding $navbar-horizontal-padding
   line-height $navbarHeight - 1.4rem
-  position relative
   a, span, img
     display inline-block
   .logo
