+++ packages/@vuepress/theme-default/components/Page.vue
@@ -62,7 +62,7 @@
 </template>
 
 <script>
-import { resolvePage, normalize, outboundRE, endingSlashRE } from '../layouts/util'
+import { resolvePage, normalize, outboundRE, endingSlashRE } from '../util'
 
 export default {
   props: ['sidebarItems'],
