+++ packages/@vuepress/theme-default/components/Sidebar.vue
@@ -23,7 +23,7 @@
 import SidebarGroup from './SidebarGroup.vue'
 import SidebarLink from './SidebarLink.vue'
 import NavLinks from './NavLinks.vue'
-import { isActive } from '../layouts/util'
+import { isActive } from '../util'
 
 export default {
   components: { SidebarGroup, SidebarLink, NavLinks },
