+++ packages/@vuepress/theme-default/layouts/Layout.vue
@@ -61,7 +61,7 @@ import Home from '../components/Home.vue'
 import Navbar from '../components/Navbar.vue'
 import Page from '../components/Page.vue'
 import Sidebar from '../components/Sidebar.vue'
-import { resolveSidebarItems } from './util'
+import { resolveSidebarItems } from '../util'
 
 export default {
   components: { Home, Page, Sidebar, Navbar },
